Use libevdev for pulling events off the fd
Includes the ability to handle SYN_DROPPED events now too.

Note: the libevdev handling is put on top of the current event loop and not
ideally integrated just yet. Instead of reading all events into the local
buffer before processing we now pull them out of libevdev into a local buffer.

Eventually this should be modified to merely pull a single event out of
libevdev and process that (libevdev has an internal buffer to keep events).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>