wl_tablet_manager: Include originating tablet with tool_added events
Not knowing which tablet a tool was first seen on doesn't matter when
the tool has a serial number, but it's a pain when it doesn't. Without
knowing the tablet that a tool originated from, we can't associate it
with said tablet until we get the first proximity_in event from the
tablet tool.

Signed-off-by: Stephen Chandler Paul <thatslyude@gmail.com>